middleware/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
numpy>=1.24.0
pandas>=2.0.0
httpx>=0.25.0
psycopg2-binary>=2.9.0
opentelemetry-api>=1.20.0
opentelemetry-instrumentation>=0.41b0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-sdk>=1.20.0
structlog>=23.1.0
psutil>=5.9.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.0.0
redis[hiredis]>=5.0.0

# CRUD and Database Integration
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
aiomysql>=0.2.0
pydantic>=2.0.0

# Telemetry and Analytics
hypothesis>=6.88.0
scipy>=1.11.0
scikit-learn>=1.3.0